Remove unused suitcase calculate_size stub
