+++ crates/gui-core/Cargo.toml
@@ -5,3 +5,5 @@ edition = "2021"
 
 [dependencies]
 egui = "0.31.1"
+psu-packer = { path = "../psu-packer" }
+serde_json = "1"
