+++ crates/psu-packer-gui/src/lib.rs
@@ -4,10 +4,12 @@ pub mod state;
 pub mod ui;
 pub mod view;
 
-pub use psu_packer::ICON_SYS_TITLE_CHAR_LIMIT;
-pub use state::{
-    MissingFileReason, MissingRequiredFile, PackerApp, ProjectRequirementStatus, SasPrefix,
+pub use gui_core::state::{
+    MissingFileReason, MissingRequiredFile, ProjectRequirementStatus, SasPrefix,
     TimestampRulesUiState, TimestampStrategy, REQUIRED_PROJECT_FILES, TIMESTAMP_FORMAT,
+    TIMESTAMP_RULES_FILE,
 };
+pub use psu_packer::ICON_SYS_TITLE_CHAR_LIMIT;
+pub use state::PackerApp;
 pub use ui::{dialogs, file_picker, pack_controls};
 pub use view::View;
