+++ crates/psu-packer-gui/src/sas_timestamps.rs
@@ -1,4 +1,3 @@
 pub(crate) use psu_packer::sas::{
-    canonical_aliases_for_category, planned_timestamp_for_folder, planned_timestamp_for_name,
-    TimestampRules,
+    planned_timestamp_for_folder, planned_timestamp_for_name, TimestampRules,
 };
