+++ crates/suitcase/src/data/virtual_file.rs
@@ -1,8 +1,1 @@
-use std::path::PathBuf;
-
-#[derive(Clone)]
-pub struct VirtualFile {
-    pub name: String,
-    pub file_path: PathBuf,
-    pub size: u64,
-}
+pub use gui_core::state::VirtualFile;
