+++ crates/suitcase/src/io/mod.rs
@@ -1,4 +1,3 @@
-pub mod calculate_size;
 pub mod export_psu;
 pub mod file_watcher;
 pub mod read_folder;
